requirements.txt
openai>=1.0
anthropic>=0.20
scipy>=1.11
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
python-dotenv>=1.0
tqdm>=4.65
